+++ src/main/java/webserver/infra/util/FileIoUtil.java
@@ -1,4 +1,4 @@
-package webserver.utils;
+package webserver.infra.util;
 
 import java.io.IOException;
 import java.net.URL;
@@ -9,7 +9,7 @@ import java.util.LinkedHashSet;
 import java.util.Optional;
 import java.util.Set;
 
-public class FileIoUtils {
+public class FileIoUtil {
     private static final Set<String> DIRECTORIES = new LinkedHashSet<>();
 
     static {
@@ -20,7 +20,7 @@ public class FileIoUtils {
     public static byte[] loadFileFromClasspath(String filePath) throws IOException {
         Optional<Path> path = Optional.empty();
         for (String dir : DIRECTORIES) {
-            Optional<Path> nowPath = Optional.ofNullable(FileIoUtils.class.getClassLoader().getResource(dir + filePath))
+            Optional<Path> nowPath = Optional.ofNullable(FileIoUtil.class.getClassLoader().getResource(dir + filePath))
                     .map(URL::getPath)
                     .map(Paths::get);
             if (nowPath.isPresent()) {
