+++ src/main/java/webserver/infra/util/IoUtil.java
@@ -1,9 +1,9 @@
-package webserver.utils;
+package webserver.infra.util;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 
-public class IOUtils {
+public class IoUtil {
     /**
      * @param BufferedReader는 Request Body를 시작하는 시점이어야
      * @param contentLength는  Request Header의 Content-Length 값이다.
