+++ src/test/java/utils/FileIoUtilTest.java
@@ -3,14 +3,14 @@ package utils;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.utils.FileIoUtils;
+import webserver.infra.util.FileIoUtil;
 
-public class FileIoUtilsTest {
-    private static final Logger log = LoggerFactory.getLogger(FileIoUtilsTest.class);
+public class FileIoUtilTest {
+    private static final Logger log = LoggerFactory.getLogger(FileIoUtilTest.class);
 
     @Test
     void loadFileFromClasspath() throws Exception {
-        byte[] body = FileIoUtils.loadFileFromClasspath("/index.html");
+        byte[] body = FileIoUtil.loadFileFromClasspath("/index.html");
         log.debug("file : {}", new String(body));
     }
 }
