+++ src/test/java/utils/IoUtilTest.java
@@ -3,13 +3,13 @@ package utils;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.utils.IOUtils;
+import webserver.infra.util.IoUtil;
 
 import java.io.BufferedReader;
 import java.io.StringReader;
 
-public class IOUtilsTest {
-    private static final Logger logger = LoggerFactory.getLogger(IOUtilsTest.class);
+public class IoUtilTest {
+    private static final Logger logger = LoggerFactory.getLogger(IoUtilTest.class);
 
     @Test
     public void readData() throws Exception {
@@ -17,6 +17,6 @@ public class IOUtilsTest {
         StringReader sr = new StringReader(data);
         BufferedReader br = new BufferedReader(sr);
 
-        logger.debug("parse body : {}", IOUtils.readData(br, data.length()));
+        logger.debug("parse body : {}", IoUtil.readData(br, data.length()));
     }
 }
